// src/core/enums/auth.enum.ts
export enum AUTH {
    EMAILINVALID = 'Invalid email format',
    EMAILEMPTY = "Email is required",
    PASSWORDNOTSTRING = "Password must be a string",
    PASSWORDLENGTH = "Password must be at least 6 characters long",
    PASSWORDEMPRY = "Password is required",
    NAMENOTSTRING = "Name must be a string",
    NAMEEMPTY = "Name is required",
    UUIDEMPTY = "Uuid is required",
    UUIDNOTSTRING = "Uuid must be a string",
    LOGOUT = "Looged out"
}